renamed io.py to avoid conflict with system
with a file named 'io.py' in the directory, some strange errors can crop
up. see:
http://stackoverflow.com/questions/6315440/how-to-solve-attributeerror-when-importing-igraph
